UAR-792 Created init methods for instance vars
We are going to use initialization methods int he construcot to encapsulate setting instance variables to default values when a new Protocol is instantiated and not loaded by OJB. We have observed that OJB seems to set the variables before the constructor is called. This causes the constructor to incorrectly set the variables to blank, empty, or default values over writing what is loaded in from the database.