UAR-687 remove import to custom class
UAR-687 Fixing import and ref to ProtocolAttachmentProtocol

UAR-687 depends on UAR-629. CustomProtocolXmlStream.java imports ProtocolAttachmentProtocol. Origianlly, in UAR-629, we created a CustomProtocolAttachmentProtocol.java but that was removed and we went back to using the org.kuali....ProtocolAttachmentProtocol.

This was fixed manually durring the last merge to development for UAR-629. However, this was never fixed in the UAR-687 branch. Since this removal was never fixed in the UAR-687 branch, we need to implement the same fix here so it will be merged correctly into the release branch and eventually master.